add endpoint to change price proposal status

+++ DTO/Responses/TopUserResponse.cs
@@ -6,7 +6,7 @@
         public int? TotalProposals { get; set; }
         public int? ApprovedProposals { get; set; }
         public int? RejectedProposals { get; set; }
-        public int? AcceptedRate { get; set; }
+        public decimal? AcceptedRate { get; set; }
         public int? Points { get; set; }
     }
 }
