+++ Data/Models/ProposalStatistic.cs
@@ -9,7 +9,7 @@
         public int? TotalProposals { get; set; }
         public int? ApprovedProposals { get; set; }
         public int? RejectedProposals { get; set; }
-        public int? AcceptedRate { get; set; }
+        public decimal? AcceptedRate { get; set; }
         public int? Points { get; set; }
         public DateTime UpdatedAt { get; set; }
     }
